Feat: More refactoring and moving classes
