+++ common/pom.xml
@@ -583,6 +583,12 @@
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
+			<groupId>com.github.retrooper</groupId>
+			<artifactId>packetevents-spigot</artifactId>
+			<version>2.7.0</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
 			<groupId>com.loohp</groupId>
 			<artifactId>InteractiveChat-Abstraction</artifactId>
 			<version>${project.version}</version>
