+++ common/src/main/java/com/loohp/interactivechat/objectholders/AsyncChatSendingExecutor.java
@@ -45,17 +45,17 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.LongSupplier;
 
-public class AsyncChatSendingExecutor implements AutoCloseable {
+public abstract class AsyncChatSendingExecutor implements AutoCloseable {
 
     private final LongSupplier executionWaitTime;
     private final long killThreadAfter;
 
     private final ReentrantLock executeLock;
-    private final Map<UUID, Queue<MessageOrderInfo>> messagesOrder;
-    private final Queue<OutboundPacket> sendingQueue;
+    public final Map<UUID, Queue<MessageOrderInfo>> messagesOrder;
+    public final Queue<OutboundPacket> sendingQueue;
     private final ThreadPoolExecutor executor;
     private final Map<Future<?>, ExecutingTaskData> executingTasks;
-    private final Map<UUID, Map<UUID, OutboundPacket>> waitingPackets;
+    public final Map<UUID, Map<UUID, OutboundPacket>> waitingPackets;
     private final Map<UUID, Long> lastSuccessfulCheck;
 
     private final List<Integer> taskIds;
@@ -100,21 +100,7 @@ public class AsyncChatSendingExecutor implements AutoCloseable {
         }
     }
 
-    public void send(PacketContainer packet, Player player, UUID id) {
-        OutboundPacket outboundPacket = new OutboundPacket(player, packet);
-        Queue<MessageOrderInfo> queue = messagesOrder.get(player.getUniqueId());
-        if (queue == null) {
-            sendingQueue.add(outboundPacket);
-        } else {
-            if (queue.stream().anyMatch(each -> each.getId().equals(id))) {
-                waitingPackets.putIfAbsent(player.getUniqueId(), new ConcurrentHashMap<>());
-                Map<UUID, OutboundPacket> waitingMap = waitingPackets.get(player.getUniqueId());
-                waitingMap.put(id, outboundPacket);
-            } else {
-                sendingQueue.add(outboundPacket);
-            }
-        }
-    }
+    public abstract void send(Object packet, Player player, UUID id);
 
     public void discard(UUID player, UUID id) {
         Queue<MessageOrderInfo> queue = messagesOrder.get(player);
@@ -198,20 +184,7 @@ public class AsyncChatSendingExecutor implements AutoCloseable {
         }, "InteractiveChat Async ChatPacket Ordered Sending Thread").start();
     }
 
-    private int packetSender() {
-        return Bukkit.getScheduler().runTaskTimer(InteractiveChat.plugin, () -> {
-            while (!sendingQueue.isEmpty()) {
-                OutboundPacket out = sendingQueue.poll();
-                try {
-                    if (out.getReciever().isOnline()) {
-                        InteractiveChat.protocolManager.sendServerPacket(out.getReciever(), out.getPacket(), false);
-                    }
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        }, 0, 1).getTaskId();
-    }
+    public abstract int packetSender();
 
     private void monitor() {
         new Thread(() -> {
@@ -285,7 +258,7 @@ public class AsyncChatSendingExecutor implements AutoCloseable {
 
     }
 
-    private static class MessageOrderInfo {
+    public static class MessageOrderInfo {
 
         private final UUID id;
         private long time;
