+++ common/src/main/resources/config.yml
@@ -174,6 +174,11 @@ Settings:
       ReplaceText: '&c{Placeholder}&r'
   Hooks:
     EcoSetLoreOnMainThread: false
+  # InteractiveChat supports both PacketEvents and ProtocolLib.
+  # You can choose which plugin should be used here.
+  # This requires a restart to apply.
+  # Possible arguments: "PacketEvents", "ProtocolLib"
+  ProtocolPlatformPlugin: "PacketEvents"
 
 Commands:
   #Whether to make commands displayed in chat clickable
